+++ packages/components/pages/f-checkout/test/visual/f-checkout-authenticated.visual.mobile.spec.js
@@ -1,10 +1,7 @@
-import forEach from 'mocha-each';
-
-const Checkout = require('../../test-utils/component-objects/f-checkout.component');
-
-let checkout;
+import Checkout from '../../test-utils/component-objects/f-checkout.component';
 
 const illegalMobileNumber = '123';
+let tests;
 
 describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => {
     const checkoutInfo = {
@@ -15,11 +12,8 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
@@ -32,8 +26,8 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
     it('should display the mandatory error messages.', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - Mandatory Errors State', 'mobile');
@@ -41,11 +35,11 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
     it('should display the "Something went wrong" error.', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'SERVER'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Something went wrong" Error', 'mobile');
@@ -53,11 +47,11 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
     it('should display the "Restaurant not taking orders" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'restaurant-not-taking-orders'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Restaurant not taking orders" Error Modal', 'mobile');
@@ -65,11 +59,11 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
     it('should display the "Additional Items Required" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'additional-items-required'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Additional Items Required" Error Modal', 'mobile');
@@ -77,9 +71,9 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - Illegal Mobile Number Error State', 'mobile');
@@ -87,11 +81,11 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Duplicate Order Warning" Modal', 'mobile');
@@ -100,11 +94,8 @@ describe('f-checkout - Collection - Authenticated - Mobile Visual Tests', () => 
 
 describe('f-checkout - Collection - Authenticated - isAsapAvailable: false Mobile Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'collection',
             isLoggedIn: true,
             isAsapAvailable: false,
@@ -127,11 +118,8 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -141,14 +129,23 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - Base State', 'mobile');
     });
 
-    forEach(['addressLine1', 'addressLocality', 'mobileNumber', 'addressPostcode'])
-    .it('should display the mandatory error messages', async field => {
-        // Act
-        await checkout.clearBlurField(field);
-        await checkout.goToPayment();
 
-        // Assert
-        await browser.percyScreenshot(`f-checkout - Delivery - Authenticated - Manadatory Error - ${field}`, 'mobile');
+    tests = [
+        'addressLine1',
+        'addressLocality',
+        'mobileNumber',
+        'addressPostcode'
+    ];
+
+    tests.forEach(field => {
+        it('should display the mandatory error messages', async () => {
+            // Act
+            await Checkout.clearBlurField(field);
+            await Checkout.goToPayment();
+
+            // Assert
+            await browser.percyScreenshot(`f-checkout - Delivery - Authenticated - Manadatory Error - ${field}`, 'mobile');
+        });
     });
 
     it('should display the illegal postcode error message', async () => {
@@ -156,9 +153,9 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
         const addressPostcode = 'TEST1A';
 
         // Act
-        await checkout.clearBlurField('addressPostcode');
-        await checkout.setFieldValue('addressPostcode', addressPostcode);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('addressPostcode');
+        await Checkout.setFieldValue('addressPostcode', addressPostcode);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - Illegal Postcode Error State', 'mobile');
@@ -166,9 +163,9 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - Illegal Mobile Number Error State', 'mobile');
@@ -176,7 +173,7 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the two notes fields if there is two noteTypes.', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             noteType: 'get-notes-config-split'
         });
@@ -187,11 +184,11 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - "Duplicate Order Warning" Modal', 'mobile');
@@ -199,11 +196,11 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the "Restaurant not taking orders" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'restaurant-not-taking-orders'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - "Restaurant not taking orders" Error Modal', 'mobile');
@@ -211,11 +208,11 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the "Additional Items Required" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'additional-items-required'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - "Additional Items Required" Error Modal', 'mobile');
@@ -224,11 +221,8 @@ describe('f-checkout - Delivery - Authenticated - Mobile Visual Tests', () => {
 
 describe('f-checkout - Delivery - Authenticated - isAsapAvailable: false Mobile Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'delivery',
             isLoggedIn: true,
             isAsapAvailable: false,
@@ -251,11 +245,8 @@ describe('f-checkout - Dine In - Authenticated - Mobile Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -267,8 +258,8 @@ describe('f-checkout - Dine In - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the mandatory error messages', async () => {
         // Act
-        ['tableIdentifier', 'mobileNumber'].forEach(async field => checkout.clearBlurField(field));
-        await checkout.goToPayment();
+        ['tableIdentifier', 'mobileNumber'].forEach(async field => Checkout.clearBlurField(field));
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine In - Authenticated - Manadatory Errors', 'mobile');
@@ -276,24 +267,21 @@ describe('f-checkout - Dine In - Authenticated - Mobile Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine In - Authenticated - Illegal Mobile Number Error State', 'mobile');
     });
 
     it('should display the "Duplicate Order Warning" modal', async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine in - Authenticated - "Duplicate Order Warning" Modal', 'mobile');