+++ packages/components/pages/f-loyalty/test/component/f-loyalty.component.spec.js
@@ -1,15 +1,12 @@
-const Loyalty = require('../../test-utils/component-objects/f-loyalty.component');
+import Loyalty from '../../test-utils/component-objects/f-loyalty.component';
 
 describe('f-loyalty component tests', () => {
-    let loyalty;
     beforeEach(async () => {
-        loyalty = new Loyalty();
-
-        await loyalty.load();
+        await Loyalty.load();
     });
 
     it('should display the f-loyalty component', async () => {
         // Assert
-        await expect(await loyalty.isComponentDisplayed()).toBe(true);
+        await expect(await Loyalty.isComponentDisplayed()).toBe(true);
     });
 });