+++ packages/components/pages/f-registration/test/accessibility/axe-accessibility.spec.js
@@ -1,17 +1,13 @@
-const { getAxeResults } = require('../../../../../../test/utils/axe-helper');
-const Registration = require('../../test-utils/component-objects/f-registration.component');
-
-const registration = new Registration();
-
+import Registration from '../../test-utils/component-objects/f-registration.component';
 
 describe('Accessibility tests', () => {
-    beforeEach(() => {
-        registration.load();
+    beforeEach(async () => {
+        await Registration.load();
     });
 
-    it('a11y - should test f-registration component WCAG compliance', () => {
+    it('a11y - should test f-registration component WCAG compliance', async () => {
         // Act
-        const axeResults = getAxeResults('f-registration');
+        const axeResults = await Registration.getAxeResults('f-registration');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);