+++ packages/components/pages/f-takeawaypay-activation/test/visual/f-takeawaypay-activation-error.visual.desktop.spec.js
@@ -1,30 +1,18 @@
-import forEach from 'mocha-each';
-
-const TakeawayPayComponent = require('../../test-utils/component-objects/f-takeawaypayActivation.component');
-
-let takeawayPayComponent;
+import TakeawayPayComponent from '../../test-utils/component-objects/f-takeawaypayActivation-error.component';
 
 describe('f-takeawaypay-activation - Error page - Desktop Visual Tests', () => {
-    beforeEach(async () => {
-        // Arrange
-        takeawayPayComponent = new TakeawayPayComponent();
-        await takeawayPayComponent
-            .withQuery('knob-Activation Status Response', '400')
-            .withQuery('knob-Employee Id', '12345')
-            .withQuery('knob-Home URL', '/home')
-            .withQuery('knob-Login URL', '/account/login')
-            .withQuery('knob-Registration URL', '/account/register');
-    });
+    const locales = [
+        'en-GB',
+        'en-AU'
+    ];
 
-    forEach(['en-GB', 'en-AU'])
-    .it('should display the component when there is an error and tenant is "%s"', async tenant => {
-        // Arrange
-        await takeawayPayComponent.withQuery('knob-Locale', tenant);
+    locales.forEach(locale => {
+        it(`should display the component when there is an error and tenant is "${locale}"`, async () => {
+            // Arrange
+            await TakeawayPayComponent.load({ locale, activationStatusResponse: '400' });
 
-        // Act
-        await takeawayPayComponent.load('error');
-
-        // Assert
-        await browser.percyScreenshot(`f-takeawaypay-activation - Error - ${tenant}`, 'desktop');
+            // Assert
+            await browser.percyScreenshot(`f-takeawaypay-activation - Error - ${locale}`, 'desktop');
+        });
     });
 });